requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
google-generativeai>=0.8
pyyaml>=6.0
python-dateutil>=2.8
